Big modification for plugin integration
Before this commit, each time that options were updated, the sanitize
methods in AdminpanelManager were called and so, updated datas has if
they were update by the user form.
So when plugins needed to update options (fields for example), the form
sanitation was done, returning an empty array.
Now, the admin panel has been updated to be able to know if the value
was send by the form or not.

Other changed :
- Add 2 hooks to be able to change fields and file_folders with plugins.
  It should be done for each option in a more pretty way
- Remove the tfi_plugins_loaded hook because it had a strange behavior
  when a plugin was activated. The load hooks was called but the files
  wasn't include, everything which should be manage with this hook
  didn't occured. And it was not the same when the tfi plugin was
  reactivated.
- Add the admin_visible option to fields and folders. This option allows
  to don't show thos fields on admin panel, really usefull for plugins.
  But the fields are not pushed on admin panel, so remember to add
  fields and folders with hooks described below